% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tax}
\alias{get_tax}
\title{Retrieve taxonomic information using taxize, and append to dataset. Note that the dataset needs a species column in [genus species] format, or separate genus and species columns.}
\usage{
get_tax(dataset, taxonomy)
}
\arguments{
\item{dataset}{a dataframe for meta-analysis, containing species names}

\item{taxonomy}{the taxonomy to retrieve}
}
\description{
Retrieve taxonomic information using taxize, and append to dataset. Note that the dataset needs a species column in [genus species] format, or separate genus and species columns.
}

