% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insects.R
\docType{data}
\name{insects}
\alias{insects}
\title{Degree days and mlower and upper development temparatures}
\format{A data frame with 1,619 rows and 8 variables}
\source{
\url{https://secure.fera.defra.gov.uk/pratique/publications.cfm}
}
\usage{
insects
}
\description{
\itemize{
  \item EADDC Degree days (celsius) above the minimum development theshold for species to completing the developmental stage.
  \item BDT Lower development threshold in degree celsius
  \item UDT Upper development threshold in degree celsius
}
}
\keyword{datasets}

