% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{plot the contribution of taxonomic levels to variance in a taxonomic analysis}
\arguments{
\item{res}{a result of class TaxMeta}

\item{taxonomy}{the taxonomy used to estimate res, must be entered as a named argument}
}
\description{
plot the contribution of taxonomic levels to variance in a taxonomic analysis
}
\author{
Philipp Neubauer
}
\references{
Neubauer.P. Minto, C, and Jensen, O.P. (in prep)
}
\seealso{
\code{\link{TaxMeta}}
}

