% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{print results a taxonomic analysis}
\arguments{
\item{res}{a result of class TaxMeta}
}
\description{
print results a taxonomic analysis
}
\author{
Philipp Neubauer
}
\references{
Neubauer.P. Minto, C, and Jensen, O.P. (in prep)
}
\seealso{
\code{\link{TaxMeta}}
}

