% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sizes.R
\docType{data}
\name{sizes}
\alias{sizes}
\title{Mammal sizes mammals}
\format{A data frame with 4,316 rows and 9 variables}
\source{
\url{http://esapubs.org/archive/ecol/E089/051/metadata.htm}
}
\usage{
sizes
}
\description{
A dataset containing mammal sizes.
}
\details{
\itemize{
  \item lmass log10 body mass
  \item cmass mass in gramm
}
}
\keyword{datasets}

