% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TaxMeta.R
\name{TaxMeta}
\alias{TaxMeta}
\title{Meta-analysis useing nested hierarchical models}
\usage{
TaxMeta(dataset, response, distribution, study_epsilon = NULL,
  taxonomy = c("species", "genus", "family", "order"), 
  save_tax = NULL, cont_cov = NULL,
  fixed_fx = NULL, random_fx = NULL, tree = NULL, scale = NULL,
  type = "full", regr_priors = NULL, loo_waic = F, n.chains = 3, n.iter = 25000,
  n.thin = 10, n.burnin = 5000, return_MCMC = F, save_mod_data=F,...)
}
\arguments{
\item{response}{the (column) name of the response variable}

\item{distribution}{a dsitribution that is supported here and implemented in JAGS (see the JAGS manual for reference). Should be one of
\itemize{
 \item \code{norm} for a normal distribution
 \item \code{gamma} for a gamma distribution
 \item \code{lnorm} for a log-normal distribution
}}

\item{study_epsilon}{the column name of individual study standard errors. This parameter is needed for a meta-analysis in the calssic sense, but is not needed if measurements are on the original scale, i.e., includes multiple datapoints for each level at the highest taxonomic resolution considered. If the latter is not the case, and study_epsilon is not available, variability at the highest taxonomic level cannot be estimated.}

\item{taxonomy}{the column names of taxonomic levels to be considered. To be given in the order from highest resolution (e.g., species) to lowest (e.g., order). For species names, please ensure that these are unique for unique species, or use a [genus species] in the species column.}

\item{save_tax}{TODO...}
\item{cont_cov}{column names of continuous covariates, if any.}

\item{fixed_fx}{column names of fixed effects, if any.}

\item{random_fx}{column names of random effects, if any.}

\item{tree}{optional, a phylogenetic tree to be combined with taxonomic effects (see Hadfield & Nakagawa, 2010).}

\item{scale}{the scale parameter for the half-cauchy hierarchical priors. The scale See Gelman 2004 for the original rationale, and ?bayou::qhalfcauchy for an R implementation of this distribution.}

\item{type}{the type of model to run, one of
\itemize{
 \item \code{full} fully hierarchical model with taxonomic variances estimated within groups at each taxonomic level. The scale for each the taxonomic model as well as the other random effects is estimated hierachically, from a global hyper-scale parameter.
 \item \code{gamma} the scale for all (taxonomic and non-taxonomic) random effects is hierarchically estimated from a vague gamma prior distribution.
 \item \code{uniform} the scale for all t(axonomic and non-taxonomic) random effects is hierarchically estimated from a uniform prior distribution.
 \item \code{fixed} the scale for all t(axonomic and non-taxonomic) random effects is fixed to \code{scale}.
}}

\item{regr_priors}{TODO...}
\item{loo_waic}{pseudo leave-one-out cross validation (see Vehtari et al., 2015) and Wanatabe (or widely applicable) information criterion (Wanatabe, 2010). NOTE that this option leads to large output objects for large datasets, that may fill all your memory (and crash your machine)!}

\item{n.chains}{number of MCMC chains (default 3)}

\item{n.iter}{number of iterations (default 20e3)}

\item{n.thin}{thinning interval of MCMC chains (default 10)}

\item{n.burnin}{number of burnin iterations to discard (default 5e3)}

\item{return_MCMC}{return MCMC chains or just summary statistics (default)}
\item{save_mod_data}{save model data that are passed to jags.model.}
}
\value{
A (list) object of class TaxMeta
}
\description{
Meta-analysis useing nested hierarchical models
}
\author{
Philipp Neubauer
}

